# frontend/public/swagger.yaml
openapi: 3.1.0
info:
  title: VRPlayground
  description: Vulnerability Research - Playground API server with various endpoints demonstrating different security vulnerabilities.
  version: 1.0.0
servers:
  - url: /api
components:
  responses:
    InternalServerError:
      description: "Internal Server Error"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: "String"
                example: "error occurred"
...
paths:
...
  /nodejs/nosql-injection:
    post:
      "tags":
        - "Vulnerability"
        - "NodeJS"
        - "NoSQL Injection"
      description: "NO-SQL Injection"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: "Text"
                  example: "admin"
                password:
                  type: string
                  description: "Password"
                  example: "admin"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    description: "Object"
                    example: [
                      {
                        "_id": "65b402a07406871e9e8b3fa3",
                        "id": 1,
                        "__v": 0,
                        "password": "admin",
                        "username": "admin"
                      }
                    ]
        '500':
          $ref: '#/components/responses/InternalServerError'
  /nodejs/path-traversal:
    get:
      "tags":
        - "Vulnerability"
        - "NodeJS"
        - "Path Traversal"
      description: "Path Traversal"
      parameters:
        - in: query
          name: path
          schema:
            type: string
            description: "Text"
            example: "public/image.png"
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                description: "Image"
                format: binary
        '500':
          $ref: '#/components/responses/InternalServerError'
  /nodejs/command-injection/argument:
    post:
      "tags":
        - "Vulnerability"
        - "NodeJS"
        - "Command Injection"
      description: "Command Injection: Argument"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                argument:
                  type: string
                  description: "Text"
                  example: "Command Argument"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    description: "String"
                    example: "Command Argument"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /nodejs/command-injection/command:
    post:
      "tags":
        - "Vulnerability"
        - "NodeJS"
        - "Command Injection"
      description: "Command Injection: Command"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                argument:
                  type: string
                  description: "Text"
                  example: "Command Argument"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    description: "String"
                    example: "Command Argument"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /nodejs/ssrf:
    post:
      "tags":
        - "Vulnerability"
        - "NodeJS"
        - "SSRF"
      description: "Server Side Request Forgery (SSRF)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: "Text"
                  example: "http://httpbin.org/get"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    description: "HTML"
                    example: '{ "args": {}, "headers": { "Host": "httpbin.org", "User-Agent": "node" }, "origin": "127.0.0.1", "url": "http://httpbin.org/get" }'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /nodejs/regex-dos:
    post:
      "tags":
        - "Vulnerability"
        - "NodeJS"
        - "Denial of Service"
      description: "Regex Denial of Service (Regex DOS)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "Text"
                  example: "realgam3@gmail.com"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    description: "Boolean"
                    example: true
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Java
...